+++ retrievalmarket/impl/lazyblockstore_test.go
@@ -14,30 +14,32 @@ import (
 )
 
 func TestLazyBlockstoreGet(t *testing.T) {
+	ctx := context.TODO()
 	b := shared_testutil.GenerateBlocksOfSize(1, 1024)[0]
 
 	ds := ds.NewMapDatastore()
 	bs := bstore.NewBlockstore(ds)
-	err := bs.Put(b)
+	err := bs.Put(ctx, b)
 	require.NoError(t, err)
 
 	lbs := newLazyBlockstore(func() (dagstore.ReadBlockstore, error) {
 		return bs, nil
 	})
 
-	blk, err := lbs.Get(b.Cid())
+	blk, err := lbs.Get(ctx, b.Cid())
 	require.NoError(t, err)
 	require.Equal(t, b, blk)
 }
 
 func TestLazyBlockstoreAllKeysChan(t *testing.T) {
+	ctx := context.TODO()
 	blks := shared_testutil.GenerateBlocksOfSize(2, 1024)
 
 	ds := ds.NewMapDatastore()
 	bs := bstore.NewBlockstore(ds)
 
 	for _, b := range blks {
-		err := bs.Put(b)
+		err := bs.Put(ctx, b)
 		require.NoError(t, err)
 	}
 
@@ -51,7 +53,7 @@ func TestLazyBlockstoreAllKeysChan(t *testing.T) {
 	var count int
 	for k := range ch {
 		count++
-		has, err := bs.Has(k)
+		has, err := bs.Has(ctx, k)
 		require.NoError(t, err)
 		require.True(t, has)
 	}
@@ -59,45 +61,48 @@ func TestLazyBlockstoreAllKeysChan(t *testing.T) {
 }
 
 func TestLazyBlockstoreHas(t *testing.T) {
+	ctx := context.TODO()
 	b := shared_testutil.GenerateBlocksOfSize(1, 1024)[0]
 
 	ds := ds.NewMapDatastore()
 	bs := bstore.NewBlockstore(ds)
-	err := bs.Put(b)
+	err := bs.Put(ctx, b)
 	require.NoError(t, err)
 
 	lbs := newLazyBlockstore(func() (dagstore.ReadBlockstore, error) {
 		return bs, nil
 	})
 
-	has, err := lbs.Has(b.Cid())
+	has, err := lbs.Has(ctx, b.Cid())
 	require.NoError(t, err)
 	require.True(t, has)
 }
 
 func TestLazyBlockstoreGetSize(t *testing.T) {
+	ctx := context.TODO()
 	b := shared_testutil.GenerateBlocksOfSize(1, 1024)[0]
 
 	ds := ds.NewMapDatastore()
 	bs := bstore.NewBlockstore(ds)
-	err := bs.Put(b)
+	err := bs.Put(ctx, b)
 	require.NoError(t, err)
 
 	lbs := newLazyBlockstore(func() (dagstore.ReadBlockstore, error) {
 		return bs, nil
 	})
 
-	sz, err := lbs.GetSize(b.Cid())
+	sz, err := lbs.GetSize(ctx, b.Cid())
 	require.NoError(t, err)
 	require.Equal(t, 1024, sz)
 }
 
 func TestLazyBlockstoreMultipleInvocations(t *testing.T) {
+	ctx := context.TODO()
 	b := shared_testutil.GenerateBlocksOfSize(1, 1024)[0]
 
 	ds := ds.NewMapDatastore()
 	bs := bstore.NewBlockstore(ds)
-	err := bs.Put(b)
+	err := bs.Put(ctx, b)
 	require.NoError(t, err)
 
 	// Count the number of times that the init function is invoked
@@ -108,10 +113,10 @@ func TestLazyBlockstoreMultipleInvocations(t *testing.T) {
 	})
 
 	// Invoke Get twice
-	_, err = lbs.Get(b.Cid())
+	_, err = lbs.Get(ctx, b.Cid())
 	require.NoError(t, err)
 
-	_, err = lbs.Get(b.Cid())
+	_, err = lbs.Get(ctx, b.Cid())
 	require.NoError(t, err)
 
 	// Verify that the init function is only invoked once